warning fix

+++ server/src/kernel/network/NetworkPacket.cpp
@@ -232,7 +232,7 @@ CAuth::doMagic(
 		drt::WorkerManager &m,
 		drt::network::PeerInfo *peer )
 {
-	if( id == -1 )
+	if( (short) id == -1 )
 	{
 		peer -> setClient();
 		peer -> setId( m.getNetwork() -> incBiggerId() );
